docs: change file name & move directory
