+++ frontend/src/constants/message.ts
@@ -1,0 +1,63 @@
+//Common
+export const ERROR_SOMETHING =
+  "Apologies, but we encountered an issue. Please try again.";
+
+export const SUCCESS_SIGNUP =
+  "Email Sent! Please check your inbox for further instructions on starting this app.";
+
+export const ERROR_FIRSTNAME = "First name is required";
+
+export const ERROR_LASTNAME = "Last name is required";
+
+export const ERROR_EMAIL = "Please enter valid email format";
+
+export const ERROR_PASSWORD =
+  "6+ characters, with at least one upper and a digit";
+
+export const ERROR_PASSWORDCONF = "Please confirm your password";
+
+export const ERROR_USER_NOTFOUND = "User not found";
+
+export const ERROR_USER_EXIST =
+  "The user already exists. Please try another email to register";
+
+//Account Page
+
+export const SUCCESS_CHANGE_ACCOUNT_INFO =
+  "Your account information has been successfully changed. If you updated your email address, please check both your current and previous email addresses for confirmation.";
+
+export const ERROR_CHANGE_ACCOUNT_INFO =
+  "An error occurred while updating your name and email. Please try again.";
+
+export const SUCCESS_RESET_PASSWORD =
+  "Email Sent! Please check your inbox for further instructions on resetting your password.";
+
+export const ERROR_RESET_PASSWORD_EMAIL_NOTHING = "Email not defined.";
+
+export const ERROR_RESET_PASSWORD_SEND_MAIL =
+  "Apologies, but we encountered an issue while attempting to send the password reset email. Please try again.";
+
+//Friend Request Page
+export const ERROR_SEND_OWN_ADDRESS =
+  "You cannot send a friend request to your email address.";
+
+export const ERROR_SEND_EXISTED_ADDRESS =
+  "You've already sent a friend request to this email.";
+
+export const ERROR_SEND_FAILED = "Failed to send an email.";
+
+export const ERROR_SELECT_FRIENDS = "Select friends from your friends list.";
+
+//Create Expense Page
+
+export const ERROR_EMPTY_DESCRIPTION = "The Description is required";
+
+export const ERROR_EMPTY_AMOUNT = "The amount is required";
+
+export const SUCCESS_CREATE_EXPENSE = "Expense has been successfully created.";
+
+//Expense Detail Page
+
+export const SUCCESS_UPDATE_EXPENSE = "Expense has been successfully updated.";
+
+export const SUCCESS_DELETE_EXPENSE = "Expense has been successfully deleted.";
