+++ frontend/src/utils/regexPatternUtils.ts
