+++ frontend/src/views/pages/LoginPage.tsx
@@ -8,14 +8,14 @@ import { useDispatch } from "react-redux";
 import { AppDispatch } from "../../store/store";
 import { login } from "../../reducer/userSlice";
 import { useState } from "react";
-import { emailRegex, passwordRegex } from "../../constants/regexPattern";
+import { emailRegex, passwordRegex } from "../../utils/regexPatternUtils";
 import { FormButton } from "../components/FormButton";
 import { Visibility, VisibilityOff } from "@mui/icons-material";
 import {
   ERROR_EMAIL,
   ERROR_PASSWORD,
   ERROR_USER_NOTFOUND,
-} from "../../utils/textUtils";
+} from "../../constants/message";
 
 const supabase = createClient(
   process.env.REACT_APP_SUPABASE_URL as string,
